moved dls_focusstack to source
