--- dls_focusstack/focus/__init__.py
@@ -1,1 +1,0 @@
-from focus_stack_lap_pyramid import *
