+++ source/dls_focusstack/config/align.ini
@@ -1,0 +1,40 @@
+# ---------------------------------- Image Alignment Configuration ----------------------------------
+# Configuration for the initial image alignment process used in the crystal matching, including the
+#  detector to be used and limits quantifying the quality of the alignment.
+
+
+# Automatically perform bulk image alignment using feature matching when a new image is selected. If
+#  disabled, the images will just be lined up by the top left corner, which is fine if you know that
+#  the images are already lined up. The alignment metric will still be calculated, indicating whether
+#  the fit is any good.
+# Possible Values: 'True' or 'False'
+# Default: True
+Perform Alignment=True
+
+# Feature detection algorithm to be used for the initial image alignment process.
+# Possible Values: 'ORB', 'SIFT', 'SURF', 'BRISK', 'MSER', 'FAST', 'STAR', 'GFTT', 'HARRIS', 'Dense', 'SimpleBlob'
+# Default: ORB
+Detector=ORB
+
+# The real size (in micrometers) represented by a single pixel in Image 1 (the formulatrix image).
+# Possible Values: Float >= 0.01
+# Default: 1.0
+Pixel Size 1 (um)=1.0
+
+# The real size (in micrometers) represented by a single pixel in Image 2 (the beamline image).
+# Possible Values: Float >= 0.01
+# Default: 1.0
+Pixel Size 2 (um)=1.0
+
+# A metric quantifying the quality of the alignment is calculated. If the metric is below this value,
+#  it is considered a good fit; if it is above, then the fit is considered poor.
+# Possible Values: Float >= 0.0
+# Default: 30.0
+Metric Limit Low=30.0
+
+# A metric quantifying the quality of the alignment is calculated. If the metric is below this value,
+#  it is considered a poor fit; if it is above, then the fit is considered to have failed totally,
+#  i.e. the images are completely dissimilar.
+# Possible Values: Float >= 0.0
+# Default: 40.0
+Metric Limit High=40.0
