+++ source/dls_focusstack/config/det_blob.ini
@@ -1,0 +1,20 @@
+# -------------------------------- SimpleBlob Detector Configuration --------------------------------
+# Implements the SimpleBlob detector
+
+
+# Determines whether or not the detector will be used. If set to false, attempting to use the
+#  detector will yield no results.
+# Possible Values: 'True' or 'False'
+# Default: True
+Enabled=True
+
+# Set the descriptor extractor type. Possible values are 'ORB', 'SURF', 'SIFT', 'BRIEF', and 'BRISK'.
+# Possible Values: 'ORB', 'SURF', 'SIFT', 'BRIEF', 'BRISK'
+# Default: BRIEF
+Extractor=BRIEF
+
+# Largest allowable keypoint distance between two features to be considered a valid match using this
+#  detector.
+# Possible Values: Integer in range [1, 100]
+# Default: 50
+Keypoint Limit=50
