+++ source/dls_focusstack/config/det_brisk.ini
@@ -1,0 +1,37 @@
+# ----------------------------------- BRISK Detector Configuration -----------------------------------
+# Implements the BRISK (Binary Robust Invariant Scalable Keypoints) keypoint detector and descriptor
+#  extractor described in 'Stefan Leutenegger, Margarita Chli and Roland Siegwart: BRISK: Binary
+#  Robust Invariant Scalable Keypoints. ICCV 2011: 2548-2555.'
+
+
+# Determines whether or not the detector will be used. If set to false, attempting to use the
+#  detector will yield no results.
+# Possible Values: 'True' or 'False'
+# Default: True
+Enabled=True
+
+# Set the descriptor extractor type. Possible values are 'ORB', 'SURF', 'SIFT', 'BRIEF', and 'BRISK'.
+# Possible Values: 'ORB', 'SURF', 'SIFT', 'BRIEF', 'BRISK'
+# Default: BRISK
+Extractor=BRISK
+
+# Largest allowable keypoint distance between two features to be considered a valid match using this
+#  detector.
+# Possible Values: Integer in range [1, 100]
+# Default: 30
+Keypoint Limit=30
+
+# FAST/AGAST detection threshold score.
+# Possible Values: Integer >= 0
+# Default: 30
+Threshold=30
+
+# Detection octaves. Use 0 to do single scale.
+# Possible Values: Integer >= 0
+# Default: 3
+Octaves=3
+
+# Apply this scale to the pattern used for sampling the neighbourhood of a keypoint.
+# Possible Values: Float >= 0.0
+# Default: 1.0
+Pattern Scale=1.0
