+++ source/dls_focusstack/config/det_sift.ini
@@ -1,0 +1,54 @@
+# ----------------------------------- SIFT Detector Configuration -----------------------------------
+# Implements the SIFT (Scale Invariant Feature Transform) keypoint detector and feature extractor
+#  detailed in the paper 'Lowe, D. G., Distinctive Image Features from Scale-Invariant Keypoints,
+#  International Journal of Computer Vision, 60, 2, pp. 91-110, 2004'. For further details see:
+#  http://docs.opencv.org/3.1.0/d5/d3c/classcv_1_1xfeatures2d_1_1SIFT.html or
+#  http://docs.opencv.org/2.4/modules/nonfree/doc/feature_detection.html
+
+
+# Determines whether or not the detector will be used. If set to false, attempting to use the
+#  detector will yield no results.
+# Possible Values: 'True' or 'False'
+# Default: True
+Enabled=True
+
+# Set the descriptor extractor type. Possible values are 'ORB', 'SURF', 'SIFT', 'BRIEF', and 'BRISK'.
+# Possible Values: 'ORB', 'SURF', 'SIFT', 'BRIEF', 'BRISK'
+# Default: SIFT
+Extractor=SIFT
+
+# Largest allowable keypoint distance between two features to be considered a valid match using this
+#  detector.
+# Possible Values: Integer in range [1, 100]
+# Default: 30
+Keypoint Limit=30
+
+# The maximum number of features to retain.
+# Possible Values: Integer >= 1
+# Default: 500
+Num Features=500
+
+# The number of layers in each octave. 3 is the value used in D. Lowe paper. The number of octaves is
+#  computed automatically from the image resolution.
+# Possible Values: Integer >= 1
+# Default: 3
+Num Octave Layers=3
+
+# The contrast threshold used to filter out weak features in semi-uniform (low-contrast) regions. The
+#  larger the threshold, the less features are produced by the detector.
+# Possible Values: Float
+# Default: 0.04
+Contrast Threshold=0.04
+
+# The threshold used to filter out edge-like features. Note that the its meaning is different from
+#  the contrastThreshold, i.e. the larger the edgeThreshold, the less features are filtered out (more
+#  features are retained).
+# Possible Values: Integer
+# Default: 10
+Edge Threshold=10
+
+# The sigma of the Gaussian applied to the input image at the octave #0. If your image is captured
+#  with a weak camera with soft lenses, you might want to reduce the number.
+# Possible Values: Float
+# Default: 1.6
+Sigma=1.6
