+++ source/dls_focusstack/config/det_surf.ini
@@ -1,0 +1,55 @@
+# ----------------------------------- SURF Detector Configuration -----------------------------------
+# Implements the SURF (Speeded Up Robust Features) keypoint detector and feature extractor detailed
+#  in the paper 'Bay, H. and Tuytelaars, T. and Van Gool, L. SURF: Speeded Up Robust Features, 9th
+#  European Conference on Computer Vision, 2006'. For further details see:
+#  http://docs.opencv.org/3.1.0/d5/df7/classcv_1_1xfeatures2d_1_1SURF.html or
+#  http://docs.opencv.org/2.4/modules/nonfree/doc/feature_detection.html
+
+
+# Determines whether or not the detector will be used. If set to false, attempting to use the
+#  detector will yield no results.
+# Possible Values: 'True' or 'False'
+# Default: True
+Enabled=True
+
+# Set the descriptor extractor type. Possible values are 'ORB', 'SURF', 'SIFT', 'BRIEF', and 'BRISK'.
+# Possible Values: 'ORB', 'SURF', 'SIFT', 'BRIEF', 'BRISK'
+# Default: SURF
+Extractor=SURF
+
+# Largest allowable keypoint distance between two features to be considered a valid match using this
+#  detector.
+# Possible Values: Integer in range [1, 100]
+# Default: 30
+Keypoint Limit=30
+
+# Threshold for the keypoint detector. Only features, whose hessian is larger than hessianThreshold
+#  are retained by the detector. Therefore, the larger the value, the less keypoints you will get. A
+#  good default value could be from 300 to 500, depending from the image contrast.
+# Possible Values: Float >= 0.0
+# Default: 100
+Hessian Threshold=100
+
+# The number of a gaussian pyramid octaves that the detector uses. It is set to 4 by default. If you
+#  want to get very large features, use the larger value. If you want just small features, decrease
+#  it.
+# Possible Values: Integer >= 0
+# Default: 4
+Num Octaves=4
+
+# The number of images within each octave of a gaussian pyramid.
+# Possible Values: Integer >= 1
+# Default: 2
+Num Octave Layers=2
+
+# Extended descriptor flag (true - use extended 128-element descriptors; false - use 64-element
+#  descriptors).
+# Possible Values: 'True' or 'False'
+# Default: True
+Extended=True
+
+# Up-right or rotated features flag (true - do not compute orientation of features; false - compute
+#  orientation).
+# Possible Values: 'True' or 'False'
+# Default: False
+Upright=False
