+++ source/dls_focusstack/config/focus_stack.ini
@@ -1,0 +1,17 @@
+
+
+# Possible Values: Integer
+# Default: 5
+Laplacian Kernel Size=5
+
+# Possible Values: Integer
+# Default: 5
+Laplacian Blur Radius=5
+
+# Possible Values: Integer
+# Default: 32
+Pyramid Minimum Size=32
+
+# Possible Values: Integer
+# Default: 12
+Number of images to stack=12
