+++ source/dls_focusstack/config/licensing.ini
@@ -1,0 +1,8 @@
+# --------------------------------- Detector Licensing Configuration ---------------------------------
+# Some detector algorithms are proprietary and are not free for commercial use.
+
+
+# Use proprietary algorithms (SIFT and SURF) in matching.
+# Possible Values: 'True' or 'False'
+# Default: True
+Non-Free Algorithms=False
