+++ source/dls_focusstack/focus/imagefft.py
