+++ source/dls_focusstack/focus/pyramid.py
