+++ source/dls_focusstack/focus/pyramid_layer.py
