+++ source/dls_focusstack/focus/sharpness_detector.py
