+++ source/dls_focusstack/gui/image_frame.py
