+++ source/dls_focusstack/gui/image_list.py
